Fixed compiler warnings (#43)
* Update hbmem.c

* Update hbrc.c

* Update rat.c

* Update substr.c

* Update browdbx.prg

* Update makefile.vc

* Update dbf1.c

* Update dbf1net.c

* Update makefile.vc

* Update compile.mak

* Update makefile.vc

* Update hbppcomp.c

* Update hbppcore.c

* Update dbstru.c

* Update decrypt.c

* Update encrypt.c

* Update eval.c

* Update field.c

* Update filter.c

* Update get.c

* Update index.c

* Update put.c

* Update rec.c

* Update replace.c

* Update sysprop.c

* Update use.c

* Update global.c

* Update hbiniseg.h

* Update compile.mak

* Update makefile.vc

* Update ChangeLog

+++ source/rtl/rat.c
@@ -169,7 +169,7 @@ HB_FUNC( RAT )
             for ( i = nStart - nLen + 1; i >= nEnd; i-- )
                 if ( memcmp( pszText->item.asString.value + i - 1, pszSub->item.asString.value, nLen ) == 0 )
                 {
-                    hb_retni( i );
+                    hb_retni( ( int ) i );
                     return;
                 }
         }
